# openapi.yaml
openapi: 3.0.1
info:
  title: Mentor-Mentee Matching API
  description: API for matching mentors and mentees in a mentoring platform
  version: 1.0.0
  contact:
    name: Mentor-Mentee Matching App
  license:
    name: MIT
servers:
  - url: http://localhost:8088/api
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /match-requests:
    post:
      operationId: createMatchRequest
      tags:
        - Match Requests
      summary: Send match request (mentee only)
      description: Send a matching request to a mentor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequestCreate'
      responses:
        '200':
          description: Match request sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRequest'
        '400':
          description: Bad request - invalid payload or mentor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match-requests/incoming:
    get:
      operationId: getIncomingMatchRequests
      tags:
        - Match Requests
      summary: Get incoming match requests (mentor only)
      description: Retrieve all match requests received by the mentor
      responses:
        '200':
          description: Incoming match requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRequest'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match-requests/outgoing:
    get:
      operationId: getOutgoingMatchRequests
      tags:
        - Match Requests
      summary: Get outgoing match requests (mentee only)
      description: Retrieve all match requests sent by the mentee
      responses:
        '200':
          description: Outgoing match requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRequestOutgoing'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match-requests/{id}/accept:
    put:
      operationId: acceptMatchRequest
      tags:
        - Match Requests
      summary: Accept match request (mentor only)
      description: Accept a specific match request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Match request ID
      responses:
        '200':
          description: Match request accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRequest'
        '404':
          description: Match request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match-requests/{id}/reject:
    put:
      operationId: rejectMatchRequest
      tags:
        - Match Requests
      summary: Reject match request (mentor only)
      description: Reject a specific match request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Match request ID
      responses:
        '200':
          description: Match request rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRequest'
        '404':
          description: Match request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match-requests/{id}:
    delete:
      operationId: cancelMatchRequest
      tags:
        - Match Requests
      summary: Cancel match request (mentee only)
      description: Cancel/delete a specific match request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Match request ID
      responses:
        '200':
          description: Match request cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRequest'
        '404':
          description: Match request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    MatchRequestCreate:
      type: object
      required:
        - mentorId
        - menteeId
        - message
      properties:
        mentorId:
          type: integer
          example: 3
        menteeId:
          type: integer
          example: 4
        message:
          type: string
          example: "멘토링 받고 싶어요!"
    MatchRequest:
      type: object
      required:
        - id
        - mentorId
        - menteeId
        - message
        - status
      properties:
        id:
          type: integer
          example: 1
        mentorId:
          type: integer
          example: 3
        menteeId:
          type: integer
          example: 4
        message:
          type: string
          example: "멘토링 받고 싶어요!"
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
          example: "pending"
    MatchRequestOutgoing:
      type: object
      required:
        - id
        - mentorId
        - menteeId
        - status
      properties:
        id:
          type: integer
          example: 11
        mentorId:
          type: integer
          example: 1
        menteeId:
          type: integer
          example: 10
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
          example: "pending"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: string
          example: "Detailed error information"
...